# Add `CompressionOptions::checksum` for enabling ZSTD checksum