--- a/db/blob/blob_file_builder.cc
+++ b/db/blob/blob_file_builder.cc
@@ -261,7 +261,7 @@ Status BlobFileBuilder::CompressBlobIfNeeded(
 
   // TODO: allow user CompressionOptions, including max_compressed_bytes_per_kb
   CompressionOptions opts;
-  CompressionContext context(blob_compression_type_, opts.level);
+  CompressionContext context(blob_compression_type_, opts);
   constexpr uint64_t sample_for_compression = 0;
 
   CompressionInfo info(opts, context, CompressionDict::GetEmptyDict(),