@cbi42 2023-08-11
> > Benchmark: I did some benchmark and did not see obvious performance impact from noisy test results
> 
> What about this one?
> 
> ```
> [Readrandom checksum=1]
> readrandom [AVG    20 runs] : 360928 (± 3579) ops/sec;    4.0 (± 0.0) MB/sec
> readrandom [MEDIAN 20 runs] : 362468 ops/sec;    4.0 MB/sec
> 
> [Readrandom checksum=0]
> readrandom [AVG    20 runs] : 380365 (± 2384) ops/sec;    4.2 (± 0.0) MB/sec
> readrandom [MEDIAN 20 runs] : 379800 ops/sec;    4.2 MB/sec
> ```
> 
> We should probably use microbench though

Updated summary with microbench result. There is some performance impact, so I also added `CompressionOptions::checksum` to optionally enable this feature.